Rename ChartIcon to RectIcon
